+++ src/models/pbaNumber.model.ts
@@ -1,0 +1,7 @@
+export interface PBANumberModel {
+  pbaNumber: string;
+  status?: string;
+  statusMessage?: string;
+  dateCreated?: string;
+  dateAccepted?: string;
+}
