+++ src/organisation/containers/organisation/organisation.component.html
@@ -1,5 +1,5 @@
 <app-hmcts-main-wrapper [title]="'Organisation'">
-  <!-- <app-prd-pba-success [response]="organisationDetails.response"></app-prd-pba-success> -->
+  <app-prd-pba-success [response]="organisationDetails.response"></app-prd-pba-success>
   <table class="govuk-table">
     <tbody class="govuk-table__body">
     <tr class="govuk-table__row">
