+++ src/organisation/containers/organisation/organisation.component.ts
@@ -1,9 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { select, Store } from '@ngrx/store';
-import { DxAddress, OrganisationContactInformation, OrganisationDetails, PBANumberModel } from 'src/models';
-import { utils } from 'src/organisation/utils';
+
+import { DxAddress, OrganisationContactInformation, OrganisationDetails, PBANumberModel } from '../../../models';
 import * as fromAuthStore from '../../../user-profile/store';
 import * as fromStore from '../../store';
+import { utils } from '../../utils';
+
 @Component({
   selector: 'app-prd-organisation-component',
   templateUrl: './organisation.component.html',
@@ -43,8 +45,8 @@ export class OrganisationComponent implements OnInit {
   }
 
   public canShowChangePbaNumbersLink(): void {
-    // this.authStore.pipe(select(fromAuthStore.getIsUserPuiFinanceManager)).subscribe((userIsPuiFinanceManager: boolean) => {
-    //   this.showChangePbaNumberLink = userIsPuiFinanceManager;
-    // });
+    this.authStore.pipe(select(fromAuthStore.getIsUserPuiFinanceManager)).subscribe((userIsPuiFinanceManager: boolean) => {
+      this.showChangePbaNumberLink = userIsPuiFinanceManager;
+    });
   }
 }
